requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
motor
langchain
langchain-openai
python-dotenv
authlib
python-jose[cryptography]
pytest
httpx
passlib[bcrypt]
celery[redis]
redis # Explicitly add redis driver if needed, celery[redis] usually includes it
